+++ index.html
@@ -8,7 +8,7 @@
     <title>Тест на определение скорости переключения внимания</title>
   </head>
   <body>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 
     <!--
     <table class="table_center">
